# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Oculide Core REST API
  description: >
    Hệ thống API RESTful và WebSocket phục vụ Giám sát phòng thực hành lập trình & Gác thi bằng AI (Oculide v1).
    Được thiết kế theo tiêu chuẩn IEEE 830-1998 Specification.
  version: 1.0.0
  contact:
    name: Oculide Development Team
    email: lean80469@gmail.com
servers:
  - url: https://api.oculide.id.vn/api/v1
    description: Production API Server
  - url: http://localhost:8000/api/v1
    description: Local Development Server
paths:
...
  /rooms:
    get:
      tags: [Rooms]
      summary: Lấy danh sách các phòng học/phòng thi
      security:
        - BearerAuth: []
      parameters:
        - name: room_type
          in: query
          schema:
            type: string
            enum: [exam, classroom]
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Danh sách phòng
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomResponse'

    post:
      tags: [Rooms]
      summary: Tạo mới một phòng thi hoặc phòng thực hành
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Phòng được tạo thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
  /rooms/public/{room_code}:
    get:
      tags: [Rooms]
      summary: Lấy thông tin công khai của phòng thi để sinh viên kiểm tra trước khi vào
      parameters:
        - name: room_code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thông tin phòng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicRoomResponse'
        '404':
          description: Không tìm thấy mã phòng
  /rooms/{room_id}:
    get:
      tags: [Rooms]
      summary: Lấy chi tiết cấu hình phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chi tiết phòng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'

    put:
      tags: [Rooms]
      summary: Chỉnh sửa cấu hình phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'

    delete:
      tags: [Rooms]
      summary: Xóa phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Đã xóa phòng
  /rooms/{room_id}/dashboard:
    get:
      tags: [Rooms]
      summary: Dữ liệu Lưới giám sát Panoptic Grid View thời gian thực
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Thông tin các sinh viên đang online, vi phạm và stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDashboardResponse'
  /rooms/{room_id}/scoreboard:
    get:
      tags: [Rooms]
      summary: Bảng điểm thời gian thực toàn bộ sinh viên trong phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Danh sách điểm bài thi của các sinh viên

  # ==========================================
  # QUESTIONS & TEST CASES
  # ==========================================
  /questions:
    post:
      tags: [Questions]
      summary: Tạo mới đề bài / câu hỏi trong phòng
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreateRequest'
      responses:
        '201':
          description: Tạo câu hỏi thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
  /questions/room/{room_id}:
    get:
      tags: [Questions]
      summary: Lấy danh sách toàn bộ câu hỏi trong phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Danh sách câu hỏi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuestionResponse'
  /questions/{question_id}:
    get:
      tags: [Questions]
      summary: Lấy chi tiết đề bài câu hỏi
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chi tiết câu hỏi và sample test cases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'

    put:
      tags: [Questions]
      summary: Cập nhật đề bài
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionUpdateRequest'
      responses:
        '200':
          description: Đã cập nhật câu hỏi

    delete:
      tags: [Questions]
      summary: Xóa câu hỏi
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Đã xóa câu hỏi
  /questions/{question_id}/test-cases:
    get:
      tags: [Questions]
      summary: Lấy danh sách test cases của câu hỏi
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Danh sách test cases

    post:
      tags: [Questions]
      summary: Thêm test case cho câu hỏi
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseCreateRequest'
      responses:
        '201':
          description: Đã tạo test case

  # ==========================================
  # SUBMISSIONS & AUTO-GRADER
  # ==========================================
...
  /chat/room/{room_id}:
    get:
      tags: [Chat]
      summary: Lấy lịch sử chat trong phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lịch sử tin nhắn
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RoomResponse:
      type: object
      properties:
        room_id:
          type: integer
        room_code:
          type: string
        room_name:
          type: string
        room_type:
          type: string
          enum: [exam, classroom]
        instructor_id:
          type: integer
        description:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        max_attempts:
          type: integer
        is_active:
          type: boolean
        livekit_room_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    PublicRoomResponse:
      type: object
      properties:
        room_id:
          type: integer
        room_code:
          type: string
        room_name:
          type: string
        room_type:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        requires_passcode:
          type: boolean
    RoomCreateRequest:
      type: object
      required: [room_code, room_name, start_time, end_time, duration_minutes]
      properties:
        room_code:
          type: string
        room_name:
          type: string
        room_type:
          type: string
          default: exam
          enum: [exam, classroom]
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        max_attempts:
          type: integer
          default: 1
        passcode:
          type: string
          nullable: true
    RoomUpdateRequest:
      type: object
      properties:
        room_name:
          type: string
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        max_attempts:
          type: integer
        is_active:
          type: boolean
        passcode:
          type: string
    RoomDashboardResponse:
      type: object
      properties:
        room_id:
          type: integer
        room_code:
          type: string
        active_count:
          type: integer
        students:
          type: array
          items:
            type: object
            properties:
              student_id:
                type: integer
              student_name:
                type: string
              status:
                type: string
              violation_count:
                type: integer
              last_snapshot:
                type: string
    QuestionResponse:
      type: object
      properties:
        question_id:
          type: integer
        room_id:
          type: integer
        question_order:
          type: integer
        question_title:
          type: string
        question_description:
          type: string
        question_type:
          type: string
          enum: [coding, multiple_choice, essay]
        programming_language:
          type: string
        max_points:
          type: number
        time_limit_minutes:
          type: integer
        memory_limit_mb:
          type: integer
    QuestionCreateRequest:
      type: object
      required: [room_id, question_order, question_title, question_description, question_type]
      properties:
        room_id:
          type: integer
        question_order:
          type: integer
        question_title:
          type: string
        question_description:
          type: string
        question_type:
          type: string
          enum: [coding, multiple_choice, essay]
        programming_language:
          type: string
        max_points:
          type: number
          default: 10.0
        time_limit_minutes:
          type: integer
        memory_limit_mb:
          type: integer
    QuestionUpdateRequest:
      type: object
      properties:
        question_title:
          type: string
        question_description:
          type: string
        max_points:
          type: number
    TestCaseCreateRequest:
      type: object
      required: [input_data, expected_output]
      properties:
        input_data:
          type: string
        expected_output:
          type: string
        is_hidden:
          type: boolean
          default: false
        points:
          type: number
          default: 0.0
...